requirements.txt
Flask==2.3.3
google-genai==0.3.0
python-dotenv==1.0.0
Pillow>=10.2.0
Werkzeug==2.3.7